% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-greatStellatedDodecahedron.R
\docType{data}
\name{greatStellatedDodecahedron}
\alias{greatStellatedDodecahedron}
\title{Great stellated dodecahedron}
\format{
A list with fields \code{vertices}, \code{faces} and \code{edges}.
}
\usage{
greatStellatedDodecahedron
}
\description{
A list representing the great stellated dodecahedron. It has
  32 vertices, 60 triangular faces and 90 edges.
}
\keyword{datasets}
