% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-NonConvexPolyhedron.R
\docType{data}
\name{NonConvexPolyhedron}
\alias{NonConvexPolyhedron}
\title{A mesh of a non-convex polyhedron}
\format{
A list with fields \code{vertices} and \code{faces}.
}
\usage{
NonConvexPolyhedron
}
\description{
A list representing a non-convex polyhedron
  with 14 vertices and 24 triangular faces.
}
\keyword{datasets}
